Move dump-thread API from agen diagnosis to profiling api
