+++ server/web-service/src/main/java/org/bithon/server/web/service/agent/api/AgentDiagnosisApi.java
@@ -20,10 +20,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotEmpty;
-import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.calcite.schema.Schema;
 import org.apache.calcite.schema.Table;
@@ -42,17 +38,11 @@ import org.apache.calcite.sql.parser.SqlParserPos;
 import org.apache.calcite.sql.util.SqlBasicVisitor;
 import org.apache.calcite.sql.validate.SqlValidatorException;
 import org.apache.calcite.util.NlsString;
-import org.bithon.agent.rpc.brpc.cmd.IJvmCommand;
-import org.bithon.component.commons.concurrency.NamedThreadFactory;
 import org.bithon.component.commons.exception.HttpMappableException;
-import org.bithon.component.commons.forbidden.SuppressForbidden;
 import org.bithon.component.commons.utils.StringUtils;
-import org.bithon.server.discovery.client.DiscoveredServiceInstance;
 import org.bithon.server.discovery.client.DiscoveredServiceInvoker;
-import org.bithon.server.discovery.declaration.controller.IAgentControllerApi;
 import org.bithon.server.web.service.WebServiceModuleEnabler;
 import org.bithon.server.web.service.agent.sql.AgentSchema;
-import org.bithon.server.web.service.agent.sql.table.AgentServiceProxyFactory;
 import org.bithon.server.web.service.agent.sql.table.IPushdownPredicateProvider;
 import org.bithon.server.web.service.common.calcite.SqlExecutionContext;
 import org.bithon.server.web.service.common.calcite.SqlExecutionEngine;
@@ -63,30 +53,16 @@ import org.bithon.server.web.service.common.output.TabSeparatedOutputFormatter;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * @author Frank Chen
@@ -100,8 +76,6 @@ public class AgentDiagnosisApi {
 
     private final SqlExecutionEngine sqlExecutionEngine;
     private final ObjectMapper objectMapper;
-    private final ApplicationContext applicationContext;
-    private final DiscoveredServiceInvoker discoveredServiceInvoker;
 
     public AgentDiagnosisApi(DiscoveredServiceInvoker discoveredServiceInvoker,
                              ObjectMapper objectMapper,
@@ -109,8 +83,6 @@ public class AgentDiagnosisApi {
         this.objectMapper = objectMapper;
         this.sqlExecutionEngine = new SqlExecutionEngine();
         this.sqlExecutionEngine.addSchema("agent", new AgentSchema(discoveredServiceInvoker, applicationContext));
-        this.applicationContext = applicationContext;
-        this.discoveredServiceInvoker = discoveredServiceInvoker;
     }
 
     @PostMapping(value = "/api/agent/query")
@@ -240,174 +212,6 @@ public class AgentDiagnosisApi {
             call.setOperand(1, SqlLiteral.createBoolean(true, new SqlParserPos(-1, -1)));
             return null;
         }
-    }
-
-    @Data
-    public static class ProfileRequest {
-        @NotEmpty
-        private String appName;
-
-        @NotEmpty
-        private String instanceName;
-
-        /**
-         * in seconds
-         */
-        @Min(3)
-        @Max(10)
-        private int interval;
-
-        /**
-         * how long the profiling should last for in seconds
-         */
-        @Max(5 * 60)
-        @Min(10)
-        private int duration;
-
-        /**
-         * Can be null or empty. if so, it defaults to cpu events.
-         * <p>
-         * Available events: cpu|alloc|nativemem|lock|cache-misses
-         */
-        private Set<String> profileEvents;
-    }
-
-    @SuppressForbidden
-    @GetMapping("/api/agent/profile/jvm")
-    public SseEmitter profile(@Valid @ModelAttribute ProfileRequest request) {
-        //
-        // Find the controller where the target instance is connected to
-        //
-        DiscoveredServiceInstance controller;
-        {
-            AtomicReference<DiscoveredServiceInstance> controllerRef = new AtomicReference<>();
-            List<DiscoveredServiceInstance> controllerList = discoveredServiceInvoker.getInstanceList(IAgentControllerApi.class);
-            CountDownLatch countDownLatch = new CountDownLatch(controllerList.size());
-            for (DiscoveredServiceInstance controllerInstance : controllerList) {
-                discoveredServiceInvoker.getExecutor()
-                                        .submit(() -> discoveredServiceInvoker.createUnicastApi(IAgentControllerApi.class, () -> controllerInstance)
-                                                                              .getAgentInstanceList(request.getAppName(), request.getInstanceName()))
-                                        .thenAccept((returning) -> {
-                                            if (!returning.isEmpty()) {
-                                                controllerRef.set(controllerInstance);
-                                            }
-                                        })
-                                        .whenComplete((ret, ex) -> countDownLatch.countDown());
-            }
-
-            try {
-                countDownLatch.await();
-            } catch (InterruptedException e) {
-                throw new RuntimeException(e);
-            }
-            if (controllerRef.get() == null) {
-                throw new HttpMappableException(HttpStatus.NOT_FOUND.value(), "No controller found for application instance [appName = %s, instanceName = %s]", request.getAppName(), request.getInstanceName());
-            }
-            controller = controllerRef.get();
-        }
-
-        //
-        // Create service proxy to agent via controller
-        //
-        AgentServiceProxyFactory agentServiceProxyFactory = new AgentServiceProxyFactory(discoveredServiceInvoker, applicationContext);
-        IJvmCommand agentJvmCommand = agentServiceProxyFactory.createUnicastProxy(IJvmCommand.class,
-                                                                                  controller,
-                                                                                  request.getAppName(),
-                                                                                  request.getInstanceName(),
-                                                                                  30_000L);
-
-        ScheduledExecutorService timer = Executors.newSingleThreadScheduledExecutor(NamedThreadFactory.daemonThreadFactory("profiling-timer"));
-        ExecutorService profilingExecutor = new ThreadPoolExecutor(1,
-                                                                   1,
-                                                                   0L,
-                                                                   TimeUnit.MILLISECONDS,
-                                                                   new LinkedBlockingQueue<>(1),
-                                                                   NamedThreadFactory.daemonThreadFactory("profiling"),
-                                                                   new ThreadPoolExecutor.DiscardPolicy());
-
-        Runnable stopProfiling = () -> {
-            timer.shutdown();
-            profilingExecutor.shutdown();
-        };
-
-        SseEmitter emitter = new SseEmitter(request.getDuration() * 1000L + 500);
-        emitter.onCompletion(stopProfiling);
-        emitter.onTimeout(stopProfiling);
-        emitter.onError((e) -> stopProfiling.run());
-
-        //
-        // Schedule a task to get information from the target instance continuously
-        //
-        final long duration = request.getDuration();
-        timer.scheduleAtFixedRate(new Runnable() {
-                                      private int elapsed = 0;
-
-                                      @Override
-                                      public void run() {
-                                          try {
-                                              emitter.send(SseEmitter.event()
-                                                                     .id(String.valueOf(elapsed))
-                                                                     .name("timer")
-                                                                     .data(Map.of("elapsed", elapsed,
-                                                                                  "remaining", duration - elapsed),
-                                                                           MediaType.APPLICATION_JSON));
-                                          } catch (IOException e) {
-                                              if (!e.getMessage().contains("Broken pipe")) {
-                                                  emitter.completeWithError(e);
-                                              }
-                                              stopProfiling.run();
-                                              return;
-                                          }
-
-                                          if (elapsed % request.getInterval() == 0) {
-
-                                              CompletableFuture.supplyAsync(agentJvmCommand::dumpThreads, profilingExecutor)
-                                                               .thenAccept(threadInfos -> {
-                                                                   SseEmitter.SseEventBuilder builder = SseEmitter.event()
-                                                                                                                  .id(String.valueOf(elapsed))
-                                                                                                                  .name("thread")
-                                                                                                                  .data(threadInfos, MediaType.APPLICATION_JSON);
-                                                                   try {
-                                                                       emitter.send(builder);
-                                                                   } catch (IllegalStateException e) {
-                                                                       if (e.getMessage() == null || !e.getMessage().contains("completed")) {
-                                                                           // the exception which is not thrown by SseEmitter
-                                                                           throw e;
-                                                                       }
-
-                                                                       // The emitter has been completed, ignore the exception
-                                                                       // This is because the above send is executed in a different thread,
-                                                                       // when it's the last round to executed, the following code might close the emitter already
-                                                                   } catch (IOException e) {
-                                                                       // Ignore the broken pipe exception which is expected when the client closes the connection
-                                                                       if (!e.getMessage().contains("Broken pipe")) {
-                                                                           emitter.completeWithError(e);
-                                                                       }
-                                                                       stopProfiling.run();
-                                                                   }
-                                                               }).exceptionally((ex) -> {
-                                                                   if (ex.getCause() != null) {
-                                                                       ex = ex.getCause();
-                                                                   }
-                                                                   stopProfiling.run();
-                                                                   emitter.completeWithError(ex);
-                                                                   log.error("Failed to get thread info", ex);
-
-                                                                   return null;
-                                                               });
-                                          }
-
-                                          if (elapsed++ >= duration) {
-                                              emitter.complete();
-                                              stopProfiling.run();
-                                          }
-                                      }
-                                  },
-                                  0,
-                                  1,
-                                  TimeUnit.SECONDS);
-
-        return emitter;
     }
 
     @ExceptionHandler({SqlValidatorException.class, SqlParseException.class})
